Moved the reworked views to a folder called new_views.
