+++ src/application/new_views/icon.ico
