--- src/application/welcome/welcome_test.py
@@ -1,2 +1,0 @@
-# Asher
-
